requirements.txt
absl-py==1.0.0
albumentations==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
asyncio==3.4.3
attrs==21.4.0
backcall==0.2.0
biopython==1.79
bleach==4.1.0
Brotli==1.0.9
cachetools==5.0.0
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
cmake==3.22.2
colorama==0.4.4
cryptography==36.0.1
cssselect2==0.5.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dlib==19.23.0
docopt==0.6.2
efficientnet-pytorch==0.6.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==2.0
fonttools==4.30.0
gast==0.5.3
gcloud==0.18.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.55.0
grpcio==1.44.0
h5py==3.6.0
hbmqtt==0.9.6
html5lib==1.1
httplib2==0.20.4
idna==3.3
imageio==2.19.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
jws==0.1.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
mat73==0.58
matlabengineforpython===R2021b
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.8.4
notebook==6.4.8
numpy==1.22.1
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
patsy==0.5.2
pdfminer.six==20211012
pdfplumber==0.6.0
pickleshare==0.7.5
Pillow==9.0.0
pretrainedmodels==0.7.4
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.0rc1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pydyf==0.1.2
pygad==2.16.3
Pygments==2.11.2
PyJWT==2.4.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyphen==0.12.0
Pyrebase4==4.5.0
pyrsistent==0.18.1
pyserial==3.5
pytesseract==0.3.9
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2021.3
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
qudida==0.0.4
radian==0.6.3
rchitect==0.3.36
rdkit-pypi==2021.9.5.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.8.0
segmentation-models-pytorch==0.2.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
stack-data==0.2.0
statsmodels==0.13.2
svgutils==0.3.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.4.12
tinycss2==1.1.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transitions==0.8.11
twilio==7.9.0
typing_extensions==4.0.1
Unidecode==1.3.4
urllib3==1.26.8
Wand==0.6.7
wcwidth==0.2.5
weasyprint==54.2
webencodings==0.5.1
websockets==10.3
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.14.0
zipp==3.7.0
zopfli==0.2.1
